+++ api/src/main/java/org/allaymc/api/world/StructureException.java
@@ -1,4 +1,4 @@
-package org.allaymc.api.utils;
+package org.allaymc.api.world;
 
 import lombok.experimental.StandardException;