+++ CHANGELOG.md
@@ -51,6 +51,7 @@ Unless otherwise specified, any version comparison below is the comparison of se
 - (API) Moved classes `BlockPropertyProcessor`, `BlockPlaceHelper` and `FortuneDropHelper` from api to server.
 - (API) Removed methods `BlockPropertyType.getProcessor()` and `BlockPropertyType.setProcessor()`.
 - (API) Refactored the permission system to add support for multiple parents in a permission group. See the commit history for details.
+- (API) Moved classes `Structure` and `StructureException` from `utils` to `world` package.
 
 ### Fixed